Fixed a crash in BOTableViewController's heightForFooterInSection method
The `footerViews` method always returns the cached array, once it has been created for the first time. This causes a crash in `tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section` if you change the number of sections of the table after the `footerViews` array has been created.

This fix resets the `footerViews` array after a new section is added